backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-docx==1.1.0
python-multipart==0.0.9
supabase==2.7.0
requests==2.32.0
Pillow==10.4.0
